+++ src/main/groovy/com/commercehub/griddle/TabularDataSource.groovy
@@ -7,4 +7,16 @@ interface TabularDataSource {
      */
     void withFile(File file, Closure tableHandler)
 
+    /**
+     * Sets the column name transformation to apply during all subsequent file processing.  The closure is passed
+     * the raw String value and should return the transformed value.  By default, a no-op transformation is used.
+     */
+    void setColumnNameTransformer(Closure<String> columnNameTransformer)
+
+    /**
+     * Sets the data cell value transformation to apply during all subsequent file processing.  The closure is passed
+     * the raw String value and should return the transformed value.  By default, a no-op transformation is used.
+     */
+    void setValueTransformer(Closure<String> valueTransformer)
+
 }