+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Change Log
 
+*   1.1.0
+    *   Support for column name transformers
+    *   Support for value transformers
+
 *   1.0.0
     *   Initial version
     *   Support for Excel (HSSF and XSSF), including custom cell mapping